+++ Content.Goobstation.Server/Speech/CheeseAccentComponent.cs
@@ -8,7 +8,7 @@
 
 using Content.Server.Speech.EntitySystems;
 
-namespace Content.Server.Speech.Components;
+namespace Content.Goobstation.Server.Speech;
 
 [RegisterComponent]
 [Access(typeof(CheeseAccentSystem))]
